ShootingTestDummy 클래스의 메소드를 DummyManager 클래스로 옮김
